refactor: 가격 타입 변경
Integer -> Long

+++ reward/src/main/java/com/nowayback/reward/application/reward/command/RewardOptionCreateCommand.java
@@ -2,7 +2,7 @@ package com.nowayback.reward.application.reward.command;
 
 public record RewardOptionCreateCommand(
         String name,
-        Integer additionalPrice,
+        Long additionalPrice,
         Integer stockQuantity,
         Boolean isRequired,
         Integer displayOrder
